--- a/bench/src/main/scala/org/bykn/bosatsu/TestBench.scala
+++ b/bench/src/main/scala/org/bykn/bosatsu/TestBench.scala
@@ -44,9 +44,10 @@ gauss$n = range($n).foldLeft(0, add)
     gauss(10)
 
   @Benchmark def bench0(): Unit = {
-    val ev = Evaluation(compiled0._1, Predef.jvmExternals)
+    val c = compiled0
+    val ev = Evaluation(c._1, Predef.jvmExternals)
     // run the evaluation
-    val res = ev.evaluateLast(compiled0._2).get._1.value
+    val res = ev.evaluateLast(c._2).get._1.value
     ()
   }
 
@@ -54,9 +55,88 @@ gauss$n = range($n).foldLeft(0, add)
     gauss(20)
 
   @Benchmark def bench1(): Unit = {
-    val ev = Evaluation(compiled1._1, Predef.jvmExternals)
+    val c = compiled1
+    val ev = Evaluation(c._1, Predef.jvmExternals)
     // run the evaluation
-    val res = ev.evaluateLast(compiled1._2).get._1.value
+    val res = ev.evaluateLast(c._2).get._1.value
+    ()
+  }
+
+  val compiled2 =
+    prepPackages(List("""
+package Euler4
+
+def operator >(a, b):
+  match a.cmp_Int(b):
+    GT: True
+    _ : False
+
+operator - = sub
+
+# given a maximum value, and a function to Option[Int], return
+# the maximum value of the function for inputs greater than 0
+# if the starting number is <= 0, we return None
+def max_of(n, fn):
+  int_loop(n, None, \i, res ->
+    next_i = i.sub(1)
+    res1 = match fn(i):
+      None: res
+      sm1 @ Some(m1):
+        match res:
+          None: sm1
+          sm @ Some(m): sm1 if m1 > m else sm
+    (next_i, res1))
+
+# return the first defined value from largest to smallest
+# of the given function, if it is defined
+def first_of(n, fn):
+  int_loop(n, None, \i, res ->
+    match fn(i):
+      None: (i - 1, None)
+      nonNone: (0, nonNone))
+
+def digit_list(n):
+  rev_list = int_loop(n, [], \n, acc ->
+    this_digit = n.mod_Int(10)
+    next_acc = [this_digit, *acc]
+    next_n = match n.div(10):
+      None:
+        # can't really happen because 10 is not zero
+        n
+      Some(next): next
+    (next_n, next_acc))
+  reverse(rev_list)
+
+def is_palindrome(lst, eq_fn):
+  (res, _) = lst.foldLeft((True, reverse(lst)), \res, item ->
+    match res:
+      (False, _): res
+      (_, []):
+        # can't really happen, the lists are the same length
+        (False, [])
+      (True, [h, *t]): (eq_fn(item, h), t))
+  res
+
+def num_is_palindrome(n):
+  digits = digit_list(n)
+  is_palindrome(digits, eq_Int)
+
+def product_palindrome(n1, n2):
+  prod = n1.times(n2)
+  Some(prod) if num_is_palindrome(prod) else None
+
+max_pal_opt = max_of(99, \n1 -> first_of(99, product_palindrome(n1)))
+
+max_pal = match max_pal_opt:
+  Some(m): m
+  None: 0
+"""), "Euler4")
+
+  @Benchmark def bench2(): Unit = {
+    val c = compiled2
+    val ev = Evaluation(c._1, Predef.jvmExternals)
+    // run the evaluation
+    val res = ev.evaluateLast(c._2).get._1.value
     ()
   }
 }