Gets almost 2x better. Not the 100x we got in #247 but still an improvement.

1. adds a benchmark of the slowest Euler problem and used that with jstack to look for cheap wins
2. the "optimization" of cleaning the Env in closures was hurting, not helping, so that was removed.
3. interning Identifiers and caching the hashCode was somewhat helpful
4. a few changes in memoization helped.

before:
```
[info] Benchmark          Mode  Cnt      Score     Error  Units   
[info] TestBench.bench0  thrpt    5  12098.270 ± 739.487  ops/s   
[info] TestBench.bench1  thrpt    5   8311.776 ± 400.025  ops/s   
[info] TestBench.bench2  thrpt    5      1.719 ±   0.115  ops/s~
```
after:
```
[info] Benchmark          Mode  Cnt      Score     Error  Units                             
[info] TestBench.bench0  thrpt    5  20597.788 ± 666.883  ops/s    
[info] TestBench.bench1  thrpt    5  13273.263 ± 306.294  ops/s    
[info] TestBench.bench2  thrpt    5      3.069 ±   0.286  ops/s
```